#164 Move Highscore to own file
<naming, organization
